Refactored data object mapping and improved database handling
Refactored the DtoMapper to improve data object mapping. Introduced better database handling mechanisms by providing proper exception handling and modularizing SQL statements in CelestialBodyDao. Enhanced test coverage by introducing a new class, CelestialBodyTest, to validate the functionality of CelestialBody objects. Also updated DB migration script to reflect the schema changes.

+++ src/main/java/com/db_course/dto/CelestialBodyDto.java
@@ -23,7 +23,7 @@ public class CelestialBodyDto {
     private boolean hasBeenExplored;
     private String radiationLevels;
     private boolean hasWater;
-    private double surfaceArea;
+    private BigDecimal surfaceArea;
     private boolean isSurfaceHard;
     private BigDecimal mass;
     private BigDecimal gravitationalFieldHeight;
